requirements.txt
pandas==1.4.1
PyAutoGUI==0.9.53
python-dotenv==0.20.0
requests==2.27.1
SQLAlchemy==1.4.32
